+++ src/components/dashboard/studio/background/BackgroundSwiper.tsx
