+++ src/components/dashboard/studio/background/BackgroundTab.tsx
