+++ src/components/dashboard/studio/background/ProductImageRequiredModal.tsx
