+++ src/components/dashboard/studio/background/SearchBar.tsx
