+++ src/components/dashboard/studio/common/TabPanel.tsx
